(feat) avoid a possible redherring when segfaulting
using something like the following frida-trace command will cause
the game to crash for some reason:
 frida-trace -H localhost -i 'libc.so.6!*' Gadget

The output will include the plumbing script that wraps tetris in
timeout, possibly causing players to spent time on something that
is not relevant to the challenge. To avoid that I added some
comments that will end up in the output of the game to steer them
away.